render/requirements.txt
Pillow>=10.0
kokoro-onnx>=0.3.6
onnxruntime>=1.17
numpy>=1.26
soundfile>=0.12
boto3>=1.34
requests>=2.31
